refactor(adapters): move translator.py into adapter submodule
